punch/requirements.txt
python-dotenv>=1.0.0,<2.0
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0,<1.0
jinja2>=3.1.0,<4.0
python-multipart>=0.0.6,<1.0
python-telegram-bot>=20.0,<22.0
apscheduler>=3.10.0,<4.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
playwright>=1.40.0
aiosqlite>=0.19.0,<1.0
httpx>=0.25.0,<1.0
pytest>=7.0
pytest-asyncio>=0.21.0
